+++ src/lantern/exporters/json.py
@@ -1,11 +1,10 @@
 import logging
-from pathlib import Path
 
 from mypy_boto3_s3 import S3Client
 
-from lantern.config import Config
 from lantern.exporters.base import ResourceExporter
 from lantern.models.record.revision import RecordRevision
+from lantern.models.site import ExportMeta
 
 
 class JsonExporter(ResourceExporter):
@@ -19,12 +18,11 @@ class JsonExporter(ResourceExporter):
     [1] https://metadata-standards.data.bas.ac.uk/standards/iso-19115-19139#json-schemas
     """
 
-    def __init__(
-        self, config: Config, logger: logging.Logger, s3: S3Client, record: RecordRevision, export_base: Path
-    ) -> None:
+    def __init__(self, logger: logging.Logger, meta: ExportMeta, s3: S3Client, record: RecordRevision) -> None:
+        export_base = meta.export_path / "records"
         export_name = f"{record.file_identifier}.json"
         super().__init__(
-            config=config, logger=logger, s3=s3, record=record, export_base=export_base, export_name=export_name
+            logger=logger, meta=meta, s3=s3, record=record, export_base=export_base, export_name=export_name
         )
 
     @property